Fix warnings when compiling luaParams

+++ src/src/luaParams.h
@@ -3,12 +3,12 @@
 #define H_LUAPARAMS
 #include "common.h"
 #include "crsf_protocol.h"
-#define LUA_DEVICE_SIZE(X) sizeof(tagLuaDeviceProperties)+strlen(X.label1)+1
-#define LUA_TEXTSELECTION_SIZE(X) sizeof(tagLuaProperties1)+strlen(X.label1)+1+strlen(X.textOption)+1+sizeof(tagLuaTextSelectionProperties)+strlen(X.label2)+1
-#define LUA_COMMAND_SIZE(X) sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaCommandProperties)+strlen(X.label2)+1
-#define LUA_UINT8_SIZE(X) sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaUint8Properties)+strlen(X.label2)+1
-#define LUA_UINT16_SIZE(X) sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaUint16Properties)+strlen(X.label2)+1
-#define LUA_STRING_SIZE(X) sizeof(tagLuaProperties1)+strlen(X.label1)+1+strlen(X.label2)+1
+#define LUA_DEVICE_SIZE(X) (uint8_t)(sizeof(tagLuaDeviceProperties)+strlen(X.label1)+1)
+#define LUA_TEXTSELECTION_SIZE(X) (uint8_t)(sizeof(tagLuaProperties1)+strlen(X.label1)+1+strlen(X.textOption)+1+sizeof(tagLuaTextSelectionProperties)+strlen(X.label2)+1)
+#define LUA_COMMAND_SIZE(X) (uint8_t)(sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaCommandProperties)+strlen(X.label2)+1)
+#define LUA_UINT8_SIZE(X) (uint8_t)(sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaUint8Properties)+strlen(X.label2)+1)
+#define LUA_UINT16_SIZE(X) (uint8_t)(sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaUint16Properties)+strlen(X.label2)+1)
+#define LUA_STRING_SIZE(X) (uint8_t)(sizeof(tagLuaProperties1)+strlen(X.label1)+1+strlen(X.label2)+1)
 
 /**we dont use this yet for OUR LUA
 #define LUA_INT8_SIZE(X) sizeof(tagLuaProperties1)+strlen(X.label1)+1+sizeof(tagLuaInt8Properties)+strlen(X.label2)+1
